hr/requirements.txt
asgiref==3.3.1
Django==3.1.4
django-extensions==3.1.0
django-js-asset==1.2.2
django-money==1.2.1
djangorestframework==3.12.2
mysql-connector-python==8.0.22
mysqlclient==2.0.2
Pillow==8.0.1
protobuf==3.14.0
py-moneyed==0.8.0
pytz==2020.4
six==1.15.0
slugify==0.0.1
sqlparse==0.4.1
